Geometry using Geogram backend

+++ src/erhe/log/erhe_log/log_glm.hpp
@@ -152,7 +152,7 @@ template <> struct fmt::formatter<glm::ivec3>
     }
 
     template <typename format_context>
-    auto format(const glm::dvec3& p, format_context& ctx) const -> decltype(ctx.out())
+    auto format(const glm::ivec3& p, format_context& ctx) const -> decltype(ctx.out())
     {
         return fmt::format_to(ctx.out(), "({}, {}, {})", p.x, p.y, p.z);
     }
@@ -171,7 +171,7 @@ template <> struct fmt::formatter<glm::ivec4>
     }
 
     template <typename format_context>
-    auto format(const glm::dvec4& p, format_context& ctx) const -> decltype(ctx.out())
+    auto format(const glm::ivec4& p, format_context& ctx) const -> decltype(ctx.out())
     {
         return fmt::format_to(ctx.out(), "({}, {}, {}, {})", p.x, p.y, p.z, p.w);
     }
